requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
timm
Pillow
numpy
python-dotenv
streamlit
requests
sqlalchemy
google-genai